refactor(schema): rename user_cosmetics to user_inventory
Rename collection, schema, model, and all TypeScript definitions
from UserCosmetics/userCosmetics to UserInventory/userInventory
to better reflect the collection's purpose.
