+++ schemas/user_inventory.js
